couple more changes
in the test file, DP algorithm was getting the correct sum, but in the
array where it shows the count of which denomination was used, it was
giving the wrong ones. I realized this is because we were starting with
the highest denom in all cases. So I put sort of a "bestIndex" variable
to store the best index, and that is the index we send to the getCoins
function.